Integrate blog to main application
Posts can now be tapped to show its content. Posts can also now be
created, edited, or deleted properly. Only an author of a post can
edit or delete his/her posts. Posts can only be created if the user
has logged in with an account.

Update the search function to be able to filter the posts by its
titles or authors.